h1/python-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-jwt-client"
version = "0.1.0"
description = "Python client for Go JWT service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "httpx==0.25.1",
    "pydantic==2.5.0",
    "python-jose[cryptography]==3.3.0"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]